src/ns_web_api/pyproject.toml
[tool.poetry]
name = "ns_web_api"
version = "0.1.0"
description = ""
authors = ["Yuchen Liu <steny138@gmail.com>"]

[tool.poetry.dependencies]
python = "3.8.5"
Flask = "^1.1.2"
Flask-Script = "^2.0.6"
Flask-SQLAlchemy = "^2.4.4"
flask-sqlacodegen = "^1.1.8"
Flask-Migrate = "^2.5.3"
Jinja2 = "^2.11.2"
uuid = "^1.30"
Flask-Bootstrap = "^3.3.7"
requests = "^2.25.1"
uWSGI = "^2.0.19"
Flask-Caching = "^1.9.0"
gunicorn = "^20.0.4"
line-bot-sdk = "^1.18.0"
google-api-python-client = "^1.12.8"
python-dotenv = "^0.15.0"
psycopg2 = "^2.8.6"
pandas = "^1.2.0"
beautifulsoup4 = "^4.9.3"
jieba = {git = "https://github.com/APCLab/jieba-tw.git", rev = "master"}
twstock = "^1.3.1"
lxml = "^4.6.2"
redis = "^3.5.3"
pytz = "^2021.1"
jsonpath = "^0.82"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
autopep8 = "^1.5.4"
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
